# implement faster rand(::StepRange{fmpz, fmpz})