This was already working via the generic implentation for arrays:
sample from `1:length(range)` and index the range with the result.
This new implementation mirrors the one for `BigInt`, and matches
its speed (so we get a speedup which can be e.g. 3x).

```julia
julia> rng = MersenneTwister();

julia> r = ZZ(1):ZZ(2)^1000;

julia> r2 = ZZ(1):ZZ(2):ZZ(2)^1000;

# master branch

julia> @btime rand($rng, $r);
  1.612 μs (30 allocations: 1.10 KiB)

julia> @btime rand($rng, $r2);
  1.604 μs (30 allocations: 1.10 KiB)

# PR
julia> @btime rand($rng, $r);
  485.733 ns (7 allocations: 200 bytes)

julia> @btime rand($rng, $r2);
  807.000 ns (11 allocations: 256 bytes)
```