@thofma 2020-09-25
How does this compare performance wise to https://github.com/thofma/Hecke.jl/blob/ffc5ae0b7aa47d0316b06e5ab58aff74cc95ea90/src/Misc/Integer.jl#L292?