@thofma 2020-09-25
Also, what about using `_fmpz_is_small`? There are already functions for those low level manipulations.