@thofma 2020-09-25
There are also functions for counting the number of limbs.