@rfourquet 2020-09-25
> Also, what about using _fmpz_is_small? 

Yes I will use that instead, thanks.

> How does this compare performance wise to https://github.com/thofma/Hecke.jl/blob/ffc5ae0b7aa47d0316b06e5ab58aff74cc95ea90/src/Misc/Integer.jl#L292?

Generally better (the one in Hecke doesn't scale well), except for very small values, e.g (Hecke vs. PR)
* `r = ZZ(1):ZZ(2)^1000` -> 3.445 μs vs. 483 ns 
* `r = ZZ(1):ZZ(2)^100` -> 772.304 vs. 414 ns
* `r = ZZ(1):ZZ(2)^50` ->  1.151 μs vs 351 ns
* `r = ZZ(1):ZZ(2)^10` -> 359 ns vs 350 ns
* `r = ZZ(1):ZZ(2)^5` -> 285 ns vs 326 ns

(these are numbers for a lenght which is a power of 2, in both cases the perf decreases a bit for a lenght a bit more than a power of 2)

Moreover, the implementation in Hecke is incorrect for ranges with a non-one step, e.g. `rand(ZZ(1):ZZ(2):ZZ(3))` can return `2`.