@rfourquet 2020-09-26
I updated according to review comments, and reviewed once more in detail and added tests, so removing "WIP".