@rfourquet 2020-10-14
There is only one build failure, which is unrelated:
> Unable to automatically install 'FLINT' from 'C:\Users\travis\.julia\packages\FLINT_jll\InNp0\Artifacts.toml'