@thofma 2020-10-16
Hm, there is a test failing in `acb_lindep`. Is this a precision issue @wbhart? Should we ignore it?