@wbhart 2020-10-16
Yes it can be ignored. Hopefully we have a ticket somewhere, but this is a heuristic method, no guarantees, so it occasionally fails.